+++ src/rjsoncons.cpp
@@ -40,91 +40,66 @@ sexp json_as(Json j, std::string as)
     }
 }
 
-// jsonpath
+// query
 
 template<class Json>
-sexp jsonpath_impl(
+Json j_query_eval(Json j, const std::string path, const std::string path_type)
+{
+    // evaluate path
+    switch(hash(path_type.c_str())) {
+    case hash("JSONpointer"): return jsonpointer::get<Json>(j, path);
+    case hash("JSONpath"): return jsonpath::json_query<Json>(j, path);
+    case hash("JMESpath"): return jmespath::search<Json>(j, path);
+    default: cpp11::stop("unknown `path_type` = '" + path_type + "'");
+    }
+}
+
+template<class Json>
+sexp j_query_impl(
     const std::string data, const std::string path,
-    const std::string as)
+    const std::string as, const std::string path_type)
 {
     Json j = Json::parse(data);
-    Json result = jsonpath::json_query<Json>(j, path);
+    Json result = j_query_eval<Json>(j, path, path_type);
     return json_as(result, as);
 }
 
 [[cpp11::register]]
-sexp cpp_jsonpath(
-    std::string data, std::string path, std::string jtype, std::string as)
+sexp cpp_j_query(
+    const std::string data, const std::string path,
+    const std::string object_names, const std::string as,
+    const std::string path_type)
 {
-    switch(hash(jtype.c_str())) {
-    case hash("asis"): return jsonpath_impl<ojson>(data, path, as);
-    case hash("sort"): return jsonpath_impl<json>(data, path, as);
-    default: cpp11::stop("unknown `object_names = '" + jtype + "'`");
-    }
-}
-
-// jmespath
-
-template<class Json>
-sexp jmespath_impl(
-    const std::string data, const std::string path, const std::string as)
-{
-    Json j = Json::parse(data);
-    Json result = jmespath::search<Json>(j, path);
-    return json_as(result, as);
-}
-
-[[cpp11::register]]
-sexp cpp_jmespath(
-    std::string data, std::string path, std::string jtype, std::string as)
-{
-    switch(hash(jtype.c_str())) {
-    case hash("asis"): return jmespath_impl<ojson>(data, path, as);
-    case hash("sort"): return jmespath_impl<json>(data, path, as);
-    default: cpp11::stop("unknown `object_names = '" + jtype + "'`");
-    }
-}
-
-// jsonpointer
-
-template<class Json>
-sexp jsonpointer_impl(
-    const std::string data, const std::string path, const std::string as)
-{
-    Json j = Json::parse(data);
-    Json result = jsonpointer::get<Json>(j, path);
-    return json_as(result, as);
-}
-
-
-[[cpp11::register]]
-sexp cpp_jsonpointer(
-    std::string data, std::string path, std::string jtype, std::string as)
-{
-    switch(hash(jtype.c_str())) {
-    case hash("asis"): return jsonpointer_impl<ojson>(data, path, as);
-    case hash("sort"): return jsonpointer_impl<json>(data, path, as);
-    default: cpp11::stop("unknown `object_names = '" + jtype + "'`");
+    switch(hash(object_names.c_str())) {
+    case hash("asis"): return j_query_impl<ojson>(data, path, as, path_type);
+    case hash("sort"): return j_query_impl<json>(data, path, as, path_type);
+    default: cpp11::stop("unknown `object_names = '" + object_names + "'`");
     }
 }
 
 // pivot
 
 template<class Json>
-sexp jsonpivot_impl(const std::string data, const std::string as)
+sexp j_pivot_impl(
+    const std::string data, const std::string path,
+    const std::string as, const std::string path_type)
 {
     Json j = Json::parse(data);
-    Json result = jsonpivot<Json>(j);
-    return json_as(result, as);
+    Json query = j_query_eval<Json>(j, path, path_type);
+    Json pivot = j_pivot<Json>(query);
+    return json_as(pivot, as);
 }
 
 [[cpp11::register]]
-sexp cpp_jsonpivot(std::string data, std::string jtype, std::string as)
+sexp cpp_j_pivot(
+    const std::string data, const std::string path,
+    const std::string object_names, const std::string as,
+    const std::string path_type)
 {
-    switch(hash(jtype.c_str())) {
-    case hash("asis"): return jsonpivot_impl<ojson>(data, as);
-    case hash("sort"): return jsonpivot_impl<json>(data, as);
-    default: cpp11::stop("unknown `object_names` = '" + jtype + "'`");
+    switch(hash(object_names.c_str())) {
+    case hash("asis"): return j_pivot_impl<ojson>(data, path, as, path_type);
+    case hash("sort"): return j_pivot_impl<json>(data, path, as, path_type);
+    default: cpp11::stop("unknown `object_names = '" + object_names + "'`");
     }
 }